Jadeite test refactoring
Refactored the opening of the project browser to be done in the test resource and reused throughout the entire suite of tests. This save quite a lot of time.

Also noticed that the categories list model was defaulting to identity based searches which failed some tests. Changed to equality search.