#386 in progress + fixed unbound method bp setting
Now get warning that you've got unsaved changes in the debugger source pane and a chance to go back. Removed copy to clipboard. This is more consistent with other browsers.

Setting breakpoints in unbound methods were found to be defective when the unbound method was updated in services on the server.

Reworked several debugger tests in the process.