ensure block in wrong spot
There's a known bug in Dolphin where it doesn't release the cursor properly in some unknown conditions. This dates back to Jade. In two key Jadeite BrowserUpdate methods there was an ensure block which was outside the Cursor showWhile:  when it should have been inside. This resulted in the inUpdate flag being left at true on occasion.