Fixes #390 two stage class compiles implemented
If the class definition string fails compilation, highlight it in the text. If a method fails compilation after a class compilation open a debugger. Not really a good way to get necessary information otherwise.

Removed RowanClassService>>isBehavior. Think this was an old experiment that is not used.

test cleanup - when closing/reopening windows look for project browsers with dirty method or class panes first.