Fixes #483 - user break & allow Jadeite to override base
In order to fix this issue, needed to override base method ``ProcessorScheduler>>userBreak:`` Dolphin (at least this version doesn't seem to have the concept of code overrides) throws a resumable error if you define a method in two packages.

To get around this annoying problem, I created a new package ``Rowan UI Base PreInstall`` whose sole job for the moment is to remove the base version of ``userBreak:``before loading package ``Rowan UI Base``. This avoids the method conflict error. Note - doing it in the pre-install package script didn't work because the conflict error comes before the preinstall script runs which is rather silly.

The fix for this problem involves sending a soft break to the session (if available) before the process is interrupted (usually with Ctrl-F12). Then after the User Interrupt message box is raised, we send a no-op server string execution to force the soft break error to be raised.