Fixes #635 Login button enablement
Set focus on the login shell before doing the query. This seems to make a difference in the button's enablement state.
Now doing a #postLogout event to make sure login shell sees that the current session is gone.
Changed the way the query was run through queryCommand:

Also, change the test to close the transcript window as a normal user would do. Avoiding this may have contributed to a false positive on the test for this issue.