Fixes #735 window position
There were three places in old Jade code that were forcibly putting the login shell, console, and project browsers to a hard-coded position. It's pretty poor style to have constants buried in the code anyway so they were removed.

We may still need to implement a "lastPosition" type of solution for all Jadeite windows depending on the needs of the issue writer.