Fixes #777 - method list source doesn't update color
Dolphin doesn't seem to handle some keys correctly. For example, if press Ctrl-X, it cuts but does not seem to run the editCut method defined in the gui. If you use the menu option (with accelerator key defined) it will run that method. In the case of #777, I  think the user was pressing the `Del` key which seems to not run smalltalk code but only windows vm code. Regardless, I added some hooks to handle changes to the view's text which is apparently different than the model's text even with updatePerCharacter turned on. Yuck.

Test updates & reverts, as well.

Also, don't update method list source if source view isModified.