Fixes #799 - toggle test selector display in sunit browser
When one test class is selected, the display will only show the column 'Test Selector'. When multiple classes are displayed, a 'Test Execution Column' will appear.

Moved some code to new packages.

Added a script that needs to (temporarily?) be run after connecting to a 3.6 server. This will connect Rowan to a local checkout and load tests.