Tests that close windows more reliable
Method sent on login was supposedly to keep methods from being stripped by the ImageStripper. Put in a never-run conditional block so the methods won't pollute the transcript with deprecation warnings. (Methods may still be needed)

After tests that close the transcript & logout, we weren't setting up the test itself. This became necessary when some tracing code run after each test that needed a valid session was put in. (In spite of the tracing code, we should probably have a valid test session anyway)

A number of tests that logout and make sure windows are closed needed retries. Encapsulated that code in #verifyWindowClosed:.

sUnit test needed selectionByIndex code updated for Dolphin which fixed a previous bug that returned an array.