Fixed autocomplete test failure - lots of other changes along the way
Registering a service for updates usually is done for a presenter in a Jadeite tool.  However, this autocomplete update is "presenter agnostic."  It was registered with a browser presenter's class list (any presenter would do). However, this created the weird problem.  If a browser wasn't open, no update would be done!  Moved the registration to the transcript which is always opened when logged in.

Clear the current BrowserUpdate upon logout. This will clear all events associated with that session's presenters.

BrowserUpdate really is a poor name choice. Should change that when come up with a better name.

Pushed postLogout changes to subclass JadeiteLoginShell.

RowanClientServices2 package is now loaded by prerequisite from a prior package (not quite sure what change prompted this - will investigate) . Commented out the explicit package load in BootJade.st.

Refactorings done in BrowserUpdate to make code more intention-revealing.

Dolphin tool discovered that Object needed to understand #isInspectorService