some accelerator keys weren't working
Seems as though this version of Dolphin has eliminated instance method #additionalAccelerators references. Implying a change to the way accelerators are setup. Moving accelerators to class side method #defaultAdditionalAccelerators resolved the problem.

Also, the tests for accelerators didn't find the problem. Enhanced the test to  simulate the accelerator key call and test the results.

Added a answering service api that returns the object of an execution, not just the printString. Only used for testing that package groups are truly from the loaded project for now.