First cut - attempted fix #927 hanging debuggers
https://github.com/GemTalk/Jadeite/issues/927

Instead of forking our own processes off to wait for an answer when a debugger opens on a remote process, we now send `deferredValue` to a block which executes asynchronously but returns synchronously.

So far, it's showing promise because the debugger tests are passing and multiple debuggers aren't hanging the system in ad-hoc testing. Need more rigorous testing though.