debugger answer method now loops safely / fixed process selection bug
https://github.com/GemTalk/Jadeite/issues/927

No need to send `loopWhile:` in JadeiteDebugger>>answer.

When selecting processes, make sure the process list in the gui is checked before making the selection. And use an `at:ifAbsent:`

Also, while waiting for an answer, if the loop is terminated, double check that the deferred value is available and return that if so.