work in progress - debugger wait code revamped again
https://github.com/GemTalk/Jadeite/issues/927
The ModalMsgLoop will not work. Blocks the main process. This wait code is more promising. It suspends the
amswering process and forks off a process to wait on the deferred value. Make sure to resume the answering process as a main process or the resume will fail. Terminate answering process if the debugger has been closed.

So far, tests (both running and debugging) & ad-hoc testing look ok.