test runs can close windows too quickly
After a window is closed, it releases itself from the server. During a test run, a session can be logged out before all the windows can close. This results in a window trying to release itself after logout. Now, we flag the session as logging out so windows know they shouldn't release themselves.