fixed extra login shells + project menu enablement problems
We were holding onto extra login shells during the tests. The culprit seems to be JadeiteLoginShell>>fillVersionList which used a sort block in a gui list. The sortBlock appeared to be hanging onto the presenter preventing garbage collection.

Removed setUp test for extra login shells as it took too long and the problem of extra login shells seems to be fixed. Left code in but commented out.

Fixed enablement project menu list enablement problem. When no project selected, reload and unload should be disabled.