JadeiteLoginTestCase>>test_gemTask failing when run
This test was failing when run but not failing when debugged. Turns out Dolphin runs test processes at priority 6 but debugs processes at priority 5. When closing windows in a gui test, this can result in views not being able to completely close because close events seem to be forked. Now, we always run the test at priority 5.