project hover help tweaked
https://github.com/GemTalk/Jadeite/issues/979

The hover help for a project makes a round trip now as the project's disk state may have changed outside Jadeite.

Removed cursor update for every round trip. The thought is that we don't want flickering cursors for every little update and the hope is that every server update is quick enough not to need it. Cursor wait's still remain for some longer operations.