Possible fix for #884 - breakpoint not being honored
https://github.com/GemTalk/Jadeite/issues/884

I believe the problem was that if the debugger was being closed or otherwise not built, the session could be nil. Then the process list was updated but the session was nil so the process was returned without updating. The debugger thought there weren't any sessions so it gave up.