Fixes #997 - gpfaults in debugger
https://github.com/GemTalk/Jadeite/issues/997

While gp faults, according to Dolphin code comments, can theoretically happen in most any external library call, we are seeing them when trying to update the method source view of the debugger.  I couldn't reproduce it but it's possible that the gui has gone away when an update for the debugger comes back to the client. There is one service update method that does an `update` of the view which was seen in the issue report. To fix the bug, we check to see of the view is open first, exiting if it's not. Then we guard against a gp fault in the udpate. Since we are likely reading corrupted memory, not writing, resuming should be ok.