Fixes #868 - wrong debugger stack frame selected
https://github.com/GemTalk/Jadeite/issues/868

Conceptually, I think you'd want the selected frame to be the first place in the stack where "your" code appears. In @LisaAlmarode 's example, that would be the frame above the "assert".  If that is true, the Jadeite code is wrong. `JadeDebugger>>stackInitialSelection` should compare a frame against all possible "error" frames and NOT compare an "error" frame against all stack frames.  With this change you'd get the frame above the first "error" frame in the stack.