Fixes #997 - problem in BrowserUpdate semaphore
https://github.com/GemTalk/Jadeite/issues/997

Although I couldn't reproduce #997, thanks to some new logging code, I was able to find a problem in the BrowserUpdate semaphore. It wasn't properly restricting access so that could be the culprit.  Marking the bug as fixed for retesting on a slow machine.

Fixing the semaphore problem created some side effects which are now fixed. Using a base Mutex instead of a custom Semaphore allows a process to enter the mutex multiple times which is desirable.

Also had to change the BrowserUpdate>>inUpdateWhile: to not grab the mutex. It is grabbed at a much lower level now. That allows a debugger to enter the mutex when an error is being debugged (debuggers use multiple processes)

* Changed some `kill` statements back to terminate so mutex's get released.

* More logging code improvements

* The conditions for opening a debugger are enhanced. We don't just rely on an invalid session when deciding whether to open up a new debugger. Use the debuggerClosed inst var, as well.