handling gpfaults incorrectly + handle bounds condition selection method
If we get a GPFault trying to refresh the breakpoint list, it means the window is gone but Dolphin isn't smart enough to realize that. So we will ignore that error. ... Also, saw a one-off where the index into a source was shorter than the source itself. Think we were in a cascade of errors but handle it nonetheless.