additional logging of freezing client + don't continue if you get a gpfault
The Dolphin documentation seems to suggest that you can continue if you get a GPFault but I'm suspicious. During a debugger update, it's possible the debugger will go away because it's multi-threaded. If we get a GPFault during an update to the method source view (which seems to be where the faults are coming from) don't just continue. Instead terminate the active process releasing semaphores. If the gui is gone, no harm is done.